Runner: Fixing main shader, fixing super bomb init

+++ materials/Materials/spritesheet.vert
@@ -1,6 +1,6 @@
 #version 100
-precision mediump int;
-precision mediump float;
+precision highp int;
+precision highp float;
 
 uniform mat4 wvpMat;
